backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-multipart==0.0.17
pydantic==2.9.2
pydantic-settings==2.6.1
sqlmodel==0.0.21
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
httpx==0.27.2
websockets==12.0
pytest==8.3.3
